tests/rename.c: print syscall() result using %ld format

+++ tests/rename.c
@@ -13,10 +13,9 @@
 int
 main(void)
 {
-	int rc = syscall(__NR_rename, OLD_FILE, NEW_FILE);
-	printf("rename(\"%s\", \"%s\") = %d %s (%m)\n",
-	       OLD_FILE, NEW_FILE, rc,
-	       errno2name());
+	long rc = syscall(__NR_rename, OLD_FILE, NEW_FILE);
+	printf("rename(\"%s\", \"%s\") = %ld %s (%m)\n",
+	       OLD_FILE, NEW_FILE, rc, errno2name());
 
 	puts("+++ exited with 0 +++");
 	return 0;
